:sparkles: :: enum을 별도로 분리
