+++ data/src/main/java/com/mpersand/data/remote/model/court/response/CourtResponse.kt
@@ -1,5 +1,8 @@
 package com.mpersand.data.remote.model.court.response
 
+import com.mpersand.data.remote.model.util.CourtType
+import com.mpersand.data.remote.model.util.DayOfWeekType
+import com.mpersand.data.remote.model.util.SportType
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
@@ -10,15 +13,3 @@ data class CourtResponse(
     val sportType: SportType,
     val courtType: CourtType
 )
-
-enum class DayOfWeekType {
-    MON, TUE, WED, THU, FRI, SAT, SUN
-}
-
-enum class SportType {
-    BASKETBALL, BADMINTON
-}
-
-enum class CourtType {
-    HALF, FULL
-}
