+++ data/src/main/java/com/mpersand/data/remote/model/util/CourtType.kt
@@ -1,0 +1,5 @@
+package com.mpersand.data.remote.model.util
+
+enum class CourtType {
+    HALF, FULL
+}
