+++ data/src/main/java/com/mpersand/data/remote/model/util/DayOfWeekType.kt
@@ -1,0 +1,5 @@
+package com.mpersand.data.remote.model.util
+
+enum class DayOfWeekType {
+    MON, TUE, WED, THU, FRI, SAT, SUN
+}
