+++ data/src/main/java/com/mpersand/data/remote/model/util/SportType.kt
@@ -1,0 +1,5 @@
+package com.mpersand.data.remote.model.util
+
+enum class SportType {
+    BASKETBALL, BADMINTON
+}
